Add configuration: templates can have materials
With these changes, aside from a 'stage' and an 'authorization', a
templates can also have materials.

I followed the 'How to add a configuration migration' guide and backed
up the previous version (75) or the xsd. Then, bumped the version to 76
and changed it in various places necessary.

This is the first step in implementing #1011